requirements.txt
elastic_apm
fastapi
numpy
pydantic
python-dotenv
sentence_transformers
starlette